Replace rate limit v1 with v2 && remove v1
