+++ store/mysql/store_conf.go
@@ -6,7 +6,7 @@ import (
 	"strconv"
 	"time"
 
-	rate "github.com/Conflux-Chain/confura/util/rate/v2"
+	"github.com/Conflux-Chain/confura/util/rate"
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
 	"gorm.io/gorm"
